backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.10
asyncpg==0.29.0
redis==5.2.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.3.0
xgboost==2.1.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
# TensorFlow and PyTorch may not support Python 3.14 yet
# If using Python 3.14, install these separately or use Python 3.11/3.12
# tensorflow>=2.16.0  # Uncomment if using Python 3.11 or 3.12
# transformers>=4.45.0  # Uncomment if using Python 3.11 or 3.12
# torch>=2.5.0  # Uncomment if using Python 3.11 or 3.12
Pillow>=10.4.0
opencv-python>=4.10.0
boto3>=1.35.0
chromadb>=0.5.0
websockets>=13.1
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
celery==5.4.0
flower==2.0.1

